pyproject.toml
[project]
name = "pixelflowlabs"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "atproto>=0.0.60",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "google-generativeai>=0.8.4",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "praw>=7.8.1",
    "pymongo>=4.12.0",
    "python-dotenv>=1.1.0",
    "schedule>=1.2.2",
    "textblob>=0.19.0",
    "transformers>=4.51.3",
]
